+++ c/common/test/rules/donotdestroyamutexwhileitislocked/test.c
@@ -1,4 +1,3 @@
-#include <stdatomic.h>
 #include <stddef.h>
 #include <threads.h>
 
@@ -20,13 +19,6 @@ int t2(void *data) {
 }
 
 int t3(void *data) {
-  mtx_t *mxl = (mtx_t *)data;
-  mtx_lock(mxl);
-  mtx_unlock(mxl);
-  return 0;
-}
-
-int t4(void *data) {
   mtx_t *mxl = (mtx_t *)data;
   mtx_lock(mxl);
   mtx_unlock(mxl);
@@ -96,37 +88,5 @@ int f4() {
   }
 
   mtx_destroy(&mxl);
-  return 0;
-}
-
-// case 5 - incorrectly waiting with a stack local variable.
-int f5() {
-  thrd_t threads[5];
-  mtx_t mxl;
-
-  mtx_init(&mxl, mtx_plain); // NON_COMPLIANT
-
-  for (size_t i = 0; i < 1; i++) {
-    thrd_create(&threads[i], t4, &mxl);
-  }
-
-  return 0;
-}
-
-// case 6 - correctly waiting with a stack local variable.
-int f6() {
-  thrd_t threads[5];
-  mtx_t mxl;
-
-  mtx_init(&mxl, mtx_plain); // COMPLIANT
-
-  for (size_t i = 0; i < 1; i++) {
-    thrd_create(&threads[i], t4, &mxl);
-  }
-
-  for (size_t i = 0; i < 1; i++) {
-    thrd_join(threads[i], 0);
-  }
-
   return 0;
 }
